changed rating datatype

+++ transform.py
@@ -28,7 +28,7 @@ def transform():
     print('====================================')
     
     # todo: change datatype 
-    
+    df["Rating"]=df["Rating"].astype(int)
     
     # todo: drop unwanted columns
     df = df.drop(['hawa'], axis=1)
